feat(task-page): reativo quando deleto o card

+++ src/app/pages/task-page/task-page.component.ts
@@ -10,7 +10,7 @@ import { TaskService } from 'src/app/tasks/task.service';
 })
 export class TaskPageComponent implements OnInit {
   tasks: Task[] = []; // Isso deve ser um array de Task
-  selectedTask !: Task 
+  selectedTask !: Task | null
   visible: boolean = false
   descricao_task : string | null = null
   botao02 = 'Excluir'
@@ -35,7 +35,7 @@ export class TaskPageComponent implements OnInit {
       this.tasks = updatedTasks
     })
   }
-  onTaskSelected(task : Task){
+  onTaskSelected(task : Task ){
     this.selectedTask = task
     
   }
@@ -51,11 +51,18 @@ export class TaskPageComponent implements OnInit {
    
     
   }
-  deleteTask(task: Task): void {
+  deleteTask(task: Task | null): void {
+    if (task ==null) {
+      return
+
+    }
+    
     this.taskService.deleteTask(task.id_task).subscribe(
       (response: any) => {
         console.log("Task excluida com sucesso", response);
         this.taskService.updateChartData()
+        this.selectedTask = null
+        
         
         
       },
@@ -65,8 +72,6 @@ export class TaskPageComponent implements OnInit {
     );
   }
 
-  handleTaskAdded() {
-    this.getTasks(); // Atualize a lista de tarefas quando uma nova tarefa for adicionada
-  }
+ 
 
 }
